# spec/openapi/p2p-wallet.yaml
openapi: 3.0.3
info:
  title: P2P Wallet API
  version: 1.1.0
  description: API для P2P кошелька

tags:
  - name: Users
    description: Операции с пользователями
  - name: Wallets
    description: Операции с кошельками

paths:
  /users/register:
    post:
      tags: [Users]
      summary: Регистрация пользователя
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/login:
    post:
      tags: [Users]
      summary: Вход в систему
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          headers:
            Set-Cookie:
              description: session_id cookie
              schema:
                type: string
                example: session_id=abc123; Path=/; HttpOnly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/{id}/logout:
    post:
      tags: [Users]
      summary: Выход из системы
      operationId: logoutUser
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: No Content
          headers:
            Set-Cookie:
              description: remove cookie
              schema:
                type: string
                example: session_id=; Path=/; HttpOnly
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /wallet:
    post:
      tags: [Wallets]
      summary: Создание нового кошелька
      operationId: createWallet
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalletRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /wallets/{user_id}:
    get:
      tags: [Wallets]
      summary: Получение кошельков пользователя
      operationId: listUserWallets
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/WalletUserId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWalletsResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: session_id

  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: ID пользователя
      schema:
        type: integer
        format: int64
        minimum: 1
    WalletUserId:
      name: user_id
      in: path
      required: true
      description: ID пользователя
      schema:
        type: integer
        format: int64
        minimum: 1

  schemas:
    RegisterRequest:
      type: object
      required:
        - login
        - password
        - name
        - last_name
      properties:
        login:
          type: string
          description: Логин (латиница)
          maxLength: 255
        password:
          type: string
          description: Пароль пользователя
          minLength: 1
        name:
          type: string
          maxLength: 50
        last_name:
          type: string
          maxLength: 50

    LoginRequest:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          maxLength: 255
        password:
          type: string
          minLength: 1

    UserResponse:
      type: object
      required:
        - id
        - login
        - name
        - last_name
        - created_at
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        name:
          type: string
        last_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    LoginResponse:
      type: object
      required:
        - id
        - login
        - name
        - last_name
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        name:
          type: string
        last_name:
          type: string

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
          example: Validation failed

    CreateWalletRequest:
      type: object
      required:
        - user_id
        - currency
      properties:
        user_id:
          type: integer
          format: int64
          minimum: 1
        currency:
          type: string
          enum: [USD, EUR]

    WalletResponse:
      type: object
      required:
        - id
        - user_id
        - currency
        - status
        - created_at
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        currency:
          type: string
          enum: [USD, EUR]
        status:
          type: string
          enum: [active, blocked]
        created_at:
          type: string
          format: date-time

    WalletItem:
      type: object
      required:
        - id
        - currency
        - total_amount
        - held_amount
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        currency:
          type: string
          enum: [USD, EUR]
        total_amount:
          type: integer
          format: int64
          description: Общий баланс в минорах
          minimum: 0
        held_amount:
          type: integer
          format: int64
          description: Удержанный баланс в минорах
          minimum: 0
        status:
          type: string
          enum: [active, blocked]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ListWalletsResponse:
      type: object
      required:
        - wallets
      properties:
        wallets:
          type: array
          items:
            $ref: '#/components/schemas/WalletItem'